Rename 'shared' module to 'i3d_classes' to better reflect what the module contains. Also create a 'utility' module, which is gonna contain small functions and definitions that are used all over to make things easier
