+++ addon/i3dio/exporter.py
@@ -16,8 +16,8 @@ from bpy_extras.io_utils import (
     axis_conversion
 )
 
-from . import shared
-from .shared import BlenderObject
+from . import i3d_classes
+from .utility import BlenderObject
 from . import xml_i3d
 from .xml_i3d import (write_attribute, add_indentations)
 from . import debugging
@@ -46,10 +46,10 @@ def export_blend_to_i3d(filepath: str, axis_forward, axis_up) -> None:
 
         # Wrap everything in a try/catch to handle addon breaking exceptions and also get them in the log file
         try:
-            i3d = shared.I3D(name=bpy.path.display_name_from_filepath(filepath),
-                             i3d_file_path=filepath,
-                             conversion_matrix=axis_conversion(to_forward=axis_forward,
-                                                               to_up=axis_up,).to_4x4())
+            i3d = i3d_classes.I3D(name=bpy.path.display_name_from_filepath(filepath),
+                                  i3d_file_path=filepath,
+                                  conversion_matrix=axis_conversion(to_forward=axis_forward,
+                                                                    to_up=axis_up,).to_4x4())
 
             export_selection = bpy.context.scene.i3dio.selection
             if export_selection == 'ALL':
@@ -77,15 +77,15 @@ def export_blend_to_i3d(filepath: str, axis_forward, axis_up) -> None:
         debugging.addon_console_handler.setLevel(debugging.addon_console_handler_default_level)
 
 
-def _export_active_scene_master_collection(i3d: shared.I3D):
+def _export_active_scene_master_collection(i3d: i3d_classes.I3D):
     pass
 
 
-def _export_active_collection(i3d: shared.I3D):
+def _export_active_collection(i3d: i3d_classes.I3D):
     pass
 
 
-def _export_active_object(i3d: shared.I3D):
+def _export_active_object(i3d: i3d_classes.I3D):
     logger.info("'Active Object' export is selected")
     if bpy.context.active_object is not None:
         _export(i3d, [bpy.context.active_object])
@@ -93,17 +93,17 @@ def _export_active_object(i3d: shared.I3D):
         logger.warning("No active object, aborting export")
 
 
-def _export_selected_objects(i3d: shared.I3D):
+def _export_selected_objects(i3d: i3d_classes.I3D):
     pass
 
 
-def _export(i3d: shared.I3D, root_objects: List[BlenderObject]):
+def _export(i3d: i3d_classes.I3D, root_objects: List[BlenderObject]):
     for blender_object in root_objects:
         _add_object_to_i3d(i3d, blender_object)
     i3d.export_to_i3d_file()
 
 
-def _add_object_to_i3d(i3d: shared.I3D, obj: BlenderObject, parent: shared.Node = None) -> None:
+def _add_object_to_i3d(i3d: i3d_classes.I3D, obj: BlenderObject, parent: i3d_classes.Node = None) -> None:
     # Collections are checked first since these are always exported in some form
     if isinstance(obj, bpy.types.Collection):
         logger.debug(f"[{obj.name}] is a 'Collection'")
@@ -141,7 +141,7 @@ def _add_object_to_i3d(i3d: shared.I3D, obj: BlenderObject, parent: shared.Node 
         logger.debug(f"[{obj.name}] no more children to process in object")
 
 
-def _process_collection_objects(i3d: shared.I3D, collection: bpy.types.Collection, parent: shared.Node):
+def _process_collection_objects(i3d: i3d_classes.I3D, collection: bpy.types.Collection, parent: i3d_classes.Node):
     """Handles adding object children of collections. Since collections stores their objects in a list named 'objects'
     instead of the 'children' list, which only contains child collections. And they need to be iterated slightly
     different"""
